+++ lib/notifications.js
@@ -1,3 +1,5 @@
+import urlBase64ToUint8Array from "@/utils/urlBase64";
+
 export async function ensurePushSubscription() {
   if (!("serviceWorker" in navigator) || !("Notification" in window)) return;
 
@@ -8,8 +10,7 @@ export async function ensurePushSubscription() {
   let subscription = await registration.pushManager.getSubscription();
 
   if (!subscription) {
-    const { publicKey } = await (await fetch("/api/push/public-key")).json();
-    const publicKeyArray = urlBase64ToUint8Array(publicKey);
+    const publicKeyArray = urlBase64ToUint8Array(process.env.NEXT_PUBLIC_VAPID_PUBLIC_KEY);
     subscription = await registration.pushManager.subscribe({
       userVisibleOnly: true,
       applicationServerKey: publicKeyArray,
@@ -21,14 +22,4 @@ export async function ensurePushSubscription() {
     headers: { "Content-Type": "application/json" },
     body: JSON.stringify(subscription),
   });
-}
-//confession: mit freundlicher Unterstützung von ChatGPT
-export function urlBase64ToUint8Array(base64) {
-  const pad = "=".repeat((4 - (base64.length % 4)) % 4);
-  const b64 = (base64 + pad).replace(/-/g, "+").replace(/_/g, "/");
-  const raw = atob(b64);
-  const out = new Uint8Array(raw.length);
-  for (let i = 0; i < raw.length; i++) out[i] = raw.charCodeAt(i);
-
-  return out;
 }
